+++ cvxpy/expressions/variables/semidefinite.py
@@ -20,13 +20,13 @@ from cvxpy.expressions.variables.variable import Variable
 from cvxpy.constraints.semi_definite import SDP
 import cvxpy.lin_ops.lin_utils as lu
 
-class Semidefinite(Variable):
+class Semidef(Variable):
     """ A semidefinite variable. """
     def __init__(self, n, name=None):
-        super(Semidefinite, self).__init__(n,n,name)
+        super(Semidef, self).__init__(n,n,name)
 
     def canonicalize(self):
         """Variable must be semidefinite and symmetric.
         """
-        obj, constr = super(Semidefinite, self).canonicalize()
+        obj, constr = super(Semidef, self).canonicalize()
         return (obj, constr + [SDP(obj)])